HBX-1570: Merge class 'org.hibernate.tool.internal.reveng.MetaAttributeXmlHelper' into 'org.hibernate.tool.internal.reveng.MetaAttributeHelper'
